Теперь танк может стрелять. Добавлены типы объектов, которые характеризуют их взаимодействия

+++ src/Renderer/VertexBufferLayout.cpp
@@ -14,7 +14,7 @@ namespace RenderEngine {
     void VertexBufferLayout::AddElementLayoutFloat(const unsigned int count, const bool isDoNormalize)
     {
         //ayoutElments.push_back({ static_cast<GLint>(count), GL_FLOAT, isDoNormalize, count * sizeof(GLfloat) });
-        unsigned int elememtSize = count * sizeof(GLfloat);
+        unsigned int elememtSize = count * static_cast<unsigned int>(sizeof(GLfloat));
         layoutElments.push_back({ static_cast<GLint>(count), GL_FLOAT, isDoNormalize, elememtSize});
         elementStride += layoutElments.back().size;
     }
